requirements/base.txt
Django==1.7.1
dj-database-url==0.3.0
django-autoslug==1.7.2
django-filter==0.9.1
django-pipeline==1.4.2
djangorestframework==3.0.2
psycopg2==2.5.4
